% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSequences.R
\name{alignSequences}
\alias{alignSequences}
\title{alignSequences}
\usage{
alignSequences(repSeqPath, refPath, out, pctid = 0.97, threads = 1)
}
\arguments{
\item{repSeqPath}{path to representative sequences}

\item{refPath}{path to reference 16S sequences}

\item{out}{output file of vsearch}

\item{pctid}{percent ID, vsearch parameter}

\item{threads}{thread number}
}
\value{
output the blast output to the specified path
}
\description{
Use vsearch to align representative sequences to reference sequence.
Will take time.
}
