% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExample.R
\name{loadExample}
\alias{loadExample}
\title{loadExample
load example input data for profileMetagenome}
\usage{
loadExample()
}
\value{
list of 16S copy number, KO copy number, BLAST result, sequence abundance
}
\description{
loadExample
load example input data for profileMetagenome
}
