% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileMetagenome.R
\name{profileMetagenome}
\alias{profileMetagenome}
\title{profileMetagenome}
\usage{
profileMetagenome(taxTable, copyNumTable, KOTable, blastRes)
}
\arguments{
\item{taxTable}{taxonomy abundance table}

\item{copyNumTable}{16S copy number table (row.names corresponds to reference tax ID)}

\item{KOTable}{precalculated KO table}

\item{blastRes}{output of alignSequences function (vsearch)
The second column must contain representative taxon ID linked to KO table.}
}
\value{
KO profile table
}
\description{
the main workflow
}
